// cil-lib/src/lib/types/operation.ts
export enum Operation {
  CREATE_ORGANIZATION = 'CREATE_ORGANIZATION',
  CREATE_SCHOOL = 'CREATE_SCHOOL',
  CREATE_CLASS = 'CREATE_CLASS',
  CREATE_USER = 'CREATE_USER',
  ADD_USERS_TO_ORGANIZATION = 'ADD_USERS_TO_ORGANIZATION',
  ADD_ORGANIZATION_ROLES_TO_USER = 'ADD_ORGANIZATION_ROLES_TO_USER',
  ADD_USERS_TO_SCHOOL = 'ADD_USERS_TO_SCHOOL',
  ADD_USERS_TO_CLASS = 'ADD_USERS_TO_CLASS',
  ADD_PROGRAMS_TO_SCHOOL = 'ADD_PROGRAMS_TO_SCHOOL',
  ADD_PROGRAMS_TO_CLASS = 'ADD_PROGRAMS_TO_CLASS',
  ADD_CLASSES_TO_SCHOOL = 'ADD_CLASSES_TO_SCHOOL',
  UNKNOWN = 'UNKNOWN',
}

export const OPERATION_ORDERING = Object.freeze([
  Operation.CREATE_ORGANIZATION,
  Operation.CREATE_SCHOOL,
  Operation.CREATE_CLASS,
  Operation.CREATE_USER,
  Operation.ADD_PROGRAMS_TO_SCHOOL,
  Operation.ADD_CLASSES_TO_SCHOOL,
  Operation.ADD_PROGRAMS_TO_CLASS,
  Operation.ADD_USERS_TO_ORGANIZATION,
  Operation.ADD_ORGANIZATION_ROLES_TO_USER,
  Operation.ADD_USERS_TO_SCHOOL,
  Operation.ADD_USERS_TO_CLASS,
]);